Move TimeProvider and this implementations to main part of project.
