+++ src/main/java/edu/iis/mto/time/timeprovider/CustomTimeProvider.java
@@ -1,4 +1,4 @@
-package edu.iis.mto.time;
+package edu.iis.mto.time.timeprovider;
 
 /**
  * Created by Wojciech Szczepaniak on 25.04.2017.
